requirements.txt
playwright
pydantic
pydantic-settings
python-dotenv
pandas
openpyxl
pdfplumber
pymupdf
tenacity
structlog
